Some code reorganisation
